+++ src/default_force.h
@@ -16,29 +16,29 @@ class DefaultForce {
   DefaultForce(const DefaultForce&) = delete;
   DefaultForce& operator=(const DefaultForce&) = delete;
 
-  void ForceBetweenSpheres(const std::array<double, 3>& ref_mass_location,
-                           double ref_diameter, double ref_iof_coefficient,
-                           const std::array<double, 3>& nb_mass_location,
-                           double nb_diameter, double nb_iof_coefficient,
-                           std::array<double, 3>* result) {
+  void ForceBetweenSpheres(const std::array<float, 3>& ref_mass_location,
+                           float ref_diameter, float ref_iof_coefficient,
+                           const std::array<float, 3>& nb_mass_location,
+                           float nb_diameter, float nb_iof_coefficient,
+                           std::array<float, 3>* result) {
     auto c1 = ref_mass_location;
-    double r1 = 0.5 * ref_diameter;
+    float r1 = 0.5 * ref_diameter;
     auto c2 = nb_mass_location;
-    double r2 = 0.5 * nb_diameter;
+    float r2 = 0.5 * nb_diameter;
     // We take virtual bigger radii to have a distant interaction, to get a
     // desired density.
-    double additional_radius =
+    float additional_radius =
         10.0 * std::min(ref_iof_coefficient, nb_iof_coefficient);
     r1 += additional_radius;
     r2 += additional_radius;
     // the 3 components of the vector c2 -> c1
-    double comp1 = c1[0] - c2[0];
-    double comp2 = c1[1] - c2[1];
-    double comp3 = c1[2] - c2[2];
-    double center_distance =
+    float comp1 = c1[0] - c2[0];
+    float comp2 = c1[1] - c2[1];
+    float comp3 = c1[2] - c2[2];
+    float center_distance =
         std::sqrt(comp1 * comp1 + comp2 * comp2 + comp3 * comp3);
     // the overlap distance (how much one penetrates in the other)
-    double delta = r1 + r2 - center_distance;
+    float delta = r1 + r2 - center_distance;
     // if no overlap : no force
     if (delta < 0) {
       *result = {0.0, 0.0, 0.0};
@@ -52,13 +52,13 @@ class DefaultForce {
       return;
     }
     // the force itself
-    double r = (r1 * r2) / (r1 + r2);
-    double gamma = 1;  // attraction coeff
-    double k = 2;      // repulsion coeff
-    double f = k * delta - gamma * std::sqrt(r * delta);
+    float r = (r1 * r2) / (r1 + r2);
+    float gamma = 1;  // attraction coeff
+    float k = 2;      // repulsion coeff
+    float f = k * delta - gamma * std::sqrt(r * delta);
 
-    double module = f / center_distance;
-    std::array<double, 3> force2on1(
+    float module = f / center_distance;
+    std::array<float, 3> force2on1(
         {module * comp1, module * comp2, module * comp3});
     *result = force2on1;
   }
